[BE/refactor] 운동 조회 시 exerciseTime만 리스트로 제공하도록 수정

See merge request s09-bigdata-recom-sub2/S09P22A403!124

+++ backend/fns/src/main/java/ssafy/fns/domain/exercise/controller/dto/SaveExerciseRequestDto.java
@@ -20,5 +20,5 @@ import lombok.Setter;
 public class SaveExerciseRequestDto {
 
     private LocalDate exerciseDate;
-    private List<Long> exerciseTimeList;
+    private List<Double> exerciseTimeList;
 }
